+++ vantage-vista/src/reference.rs
@@ -24,8 +24,10 @@ impl Reference {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
 pub enum ReferenceKind {
+    #[default]
     HasOne,
     HasMany,
     HasForeign,